+++ packages/next-common/utils/viewfuncs/index.js
@@ -3,6 +3,13 @@ import Chains from "../consts/chains";
 import { BalanceDecimals } from "../constants";
 import { detailPageCategory } from "../consts/business/category";
 import isMoonChain from "../isMoonChain";
+import { getMotionId } from "next-common/utils/motion";
+import { getTitle } from "next-common/utils/post";
+import {
+  advisoryCommitteeBaseUrl,
+  childBountyBaseUrl,
+} from "next-common/utils/postBaseUrl";
+import { getPostLastActivityAt } from "next-common/utils/viewfuncs/postUpdatedTime";
 
 export function toApiType(type) {
   // Open Gov
@@ -148,3 +155,39 @@ const isBase58 = (value) => /^[A-HJ-NP-Za-km-z1-9]*$/.test(value);
 export const isPolkadotAddress = (address) => {
   return [46, 47, 48, 49].includes(address?.length) && isBase58(address);
 };
+
+export const toFinancialMotionsListItem = (item) => {
+  return {
+    ...item,
+    index: item.motionIndex,
+    title: getTitle(item),
+    author: item.author,
+    address: item.proposer,
+    status: item.state ?? "Unknown",
+    detailLink: `/financial-council/motions/${getMotionId(item)}`,
+    time: getPostLastActivityAt(item),
+  };
+};
+
+export const toAdvisoryMotionsListItem = (item) => ({
+  ...item,
+  index: item.motionIndex,
+  title: getTitle(item),
+  author: item.author,
+  address: item.proposer,
+  status: item.state ?? "Unknown",
+  detailLink: `${advisoryCommitteeBaseUrl}/${getMotionId(item)}`,
+  time: getPostLastActivityAt(item),
+});
+
+export const toTreasuryChildBountyListItem = (item) => ({
+  ...item,
+  title: getTitle(item),
+  author: item.author,
+  address: item.proposer,
+  status: item.state ?? "Unknown",
+  time: getPostLastActivityAt(item),
+  value: item.onchainData.value,
+  detailLink: `${childBountyBaseUrl}/${item.index}`,
+  parentIndex: item.parentBountyId,
+});
